Add styles for houses slider
