--- src/components/House.jsx
@@ -1,9 +1,0 @@
-const House = () => {
-    return (
-        <div>
-            House
-        </div>
-    )
-}
-
-export default House;
